app/requirements.txt
mysql-connector-python
flask  # or any other dependencies your Python application needs